locale.c: Omit final ';' in LC_ALL composite locale
When locales for categories differ, LC_ALL needs a way to express the
individual components.  Prior to this commit, the aggregation ended with
a semi-colon, which is unnecessary.  This commit removes that.

It also moves a loop counter declaration into the loops, now that we
have C99, and avoids an extra strlen() call.