pp_ctl.c - dedupe @INC entries, especially hooks.
Allowing a hook to execute more than once opens up a bunch of failure
modes that are difficult to debug and potentially fatal. Simply skipping
anything seen twice makes all the pain go away. It also might save some
time in terms of not checking the same dir twice, but that is a side
benefit.

Currently this breaks a test, but that test is kinda broken anyway
and will be fixed in a follow up patch.