Conform '--podpath'
The previous documentation for '--podpath' in bin/pod2html was:

    "Specify the base directory for finding library pods. Default is
    the current working directory."

This could have been (mis-)interpreted to suggest that such a "base
directory" not have a necessary relationship to podroot.  Inspection of
the code indicates that any directory mentioned in podpath which is not
underneath podroot will simply be ignored.